services atualizados

+++ src/user/user.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
@@ -12,29 +16,76 @@ export class UserService {
     @InjectRepository(User) private userRepository: Repository<User>,
   ) {}
 
-  async getAllUsers() {
+  async getAllUsers(): Promise<User[]> {
     const users = await this.userRepository.find();
     return users;
   }
 
-  async getUserById(userId: number) {
-    const user = this.userRepository.findOneBy({ id: userId });
+  async getUserById(userId: number): Promise<User> {
+    const user = await this.userRepository.findOneBy({ id: userId });
+
+    if (!user) {
+      throw new NotFoundException(
+        `Usuário com id ${userId} não foi encontrado.`,
+      );
+    }
+
     return user;
   }
 
-  async createUser(dto: CreateUserDto) {
+  async createUser(dto: CreateUserDto): Promise<User> {
+    const existingUser = await this.userRepository.findOneBy({
+      username: dto.username,
+    });
+
+    if (existingUser) {
+      throw new BadRequestException(
+        `Já existe um usuário com o nome ${dto.username}.`,
+      );
+    }
+
     const newUser = this.userRepository.create(dto);
     await this.userRepository.save(newUser);
+
     return newUser;
   }
 
-  async updateUser(userId: number, dto: UpdateUserDto) {
+  async updateUser(userId: number, dto: UpdateUserDto): Promise<User> {
+    const existingUser = await this.userRepository.findOneBy({
+      username: dto.username,
+    });
+
+    if (existingUser) {
+      throw new BadRequestException(
+        `Já existe um usuário com o nome ${dto.username}.`,
+      );
+    }
+
     await this.userRepository.update(userId, dto);
     const user = await this.userRepository.findOneBy({ id: userId });
+
+    if (!user) {
+      throw new NotFoundException(
+        `Usuaŕio com id ${userId} não foi encontrado.`,
+      );
+    }
+
     return user;
   }
 
   async deleteUser(userId: number) {
-    return await this.userRepository.delete(userId);
+    const user = await this.userRepository.findOneBy({ id: userId });
+
+    if (!user) {
+      throw new NotFoundException(
+        `Usuário com id ${userId} não foi encontrado.`,
+      );
+    }
+
+    await this.userRepository.delete(userId);
+
+    return {
+      message: `Usuário com id ${userId} foi deletado com sucesso.`,
+    };
   }
 }
